feat: modified options interface
